+++ app/src/main/java/com/brandonjamesyoung/levelup/di/IconModule.kt
@@ -1,7 +1,7 @@
 package com.brandonjamesyoung.levelup.di
 
 import android.content.Context
-import com.brandonjamesyoung.levelup.utility.QuestGridCreator
+import com.brandonjamesyoung.levelup.utility.IconGridCreator
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -11,10 +11,10 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-class CardModule {
+class IconModule {
     @Singleton
     @Provides
-    fun provideCardGridCreator(@ApplicationContext context: Context): QuestGridCreator {
-        return QuestGridCreator(context)
+    fun provideIconGridCreator(@ApplicationContext context: Context): IconGridCreator {
+        return IconGridCreator(context)
     }
 }